ai_microservices/coffee_chatbot_backend/requirements.txt
fastapi
google-genai
httpx
matplotlib
networkx
nltk
numpy
packaging
pandas
psycopg2-binary
pydantic
python-dotenv
scikit-learn
setuptools
six
sqlalchemy
uvicorn
wheel
torch
sentence-transformers
